LayoutTests/fast/device-orientation/device-orientation-request-permission-user-gesture-expected.txt

CONSOLE MESSAGE: NotAllowedError: Requesting device orientation access requires a user gesture to prompt
Tests that DeviceOrientationEvent.requestPermission() requires a user gesture.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS e.name is "NotAllowedError"
PASS successfullyParsed is true

TEST COMPLETE
